+++ sandwich/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased changes
 
 * Add primed versions of createProcessWithLogging etc. with customizable log level
+* Add `Test.Sandwich.Util.Process` with `gracefullyStopProcess` and `gracefullyWaitForProcess` (and remove these from an internal `sandwich-webdriver` module).
 
 ## 0.2.1.0